+++ theme/onyx/ThemeState.qml
@@ -3,12 +3,8 @@ import QtQuick
 QtObject {
     id: themeState
 
-    FontLoader {
-        id: outfitFont
-        source: "font/Outfit-Black.ttf"
-    }
-
     property real s: 1
+    property string fontFamily: "Sans Serif"
 
     readonly property bool isPreview: typeof sddm === "undefined" || sddm.hostName === undefined
 
@@ -29,7 +25,5 @@ QtObject {
     readonly property color orbitalTickColor: "#FFFFFF"
     readonly property color orbitalTextColor: "#CCCCCC"
     readonly property color errorColor: "#FF4444"
-    readonly property string fontFamily: outfitFont.status === FontLoader.Ready
-                                          ? outfitFont.name
-                                          : "Sans Serif"
+
 }